+++ src/lapl.h
@@ -82,7 +82,7 @@ public:
 };
 
 extern "C" {
-	void SPAPI sp_lapl(double * out, const double * in, long long * nlat, long long * nlon);
+	void SPAPI sp_lapl(double * out, const double * in, long * nlat, long * nlon);
 }
 
 #endif /* LAPL_H */
